Refactor the timer utility into its own library.
