+++ centrallix-lib/include/timer.h
@@ -1,0 +1,36 @@
+#ifndef TIMER_H
+#define TIMER_H
+
+/************************************************************************/
+/* Centrallix Application Server System                                 */
+/* Centrallix Base Library                                              */
+/*                                                                      */
+/* Copyright (C) 1998-2026 LightSys Technology Services, Inc.           */
+/*                                                                      */
+/* You may use these files and this library under the terms of the      */
+/* GNU Lesser General Public License, Version 2.1, contained in the     */
+/* included file "COPYING".                                             */
+/*                                                                      */
+/* Module:      timer.c, timer.h                                        */
+/* Author:      Israel Fuller                                           */
+/* Date:        October 13, 2025                                        */
+/* Description: A simple timer utility, intended for benchmarking code  */
+/*              performance in wall time.                               */
+/************************************************************************/
+
+typedef struct
+    {
+    double start, total;
+    }
+    Timer, *pTimer;
+
+pTimer timer_init(pTimer timer);
+pTimer timer_new(void);
+pTimer timer_start(pTimer timer);
+pTimer timer_stop(pTimer timer);
+double timer_get(pTimer timer);
+pTimer timer_reset(pTimer timer);
+void timer_de_init(pTimer timer);
+void timer_free(pTimer timer);
+
+#endif /* TIMER_H */
