+++ centrallix-lib/src/util.c
@@ -13,7 +13,6 @@
 /* Date:	May 26, 2011 and October 13, 2025 (respectively)	*/
 /* Description:	Collection of utilities including:			*/
 /* 		- Utilities for parsing numbers.			*/
-/* 		- The timer utility for benchmarking code.		*/
 /* 		- snprint_bytes() for formatting a byte count.		*/
 /* 		- snprint_commas_llu() for formatting large numbers.	*/
 /* 		- fprint_mem() for printing memory stats.		*/
@@ -27,7 +26,6 @@
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <time.h>
 #include <unistd.h>
 
 #include "newmalloc.h"
@@ -217,119 +215,6 @@ fprint_mem(FILE* out)
 	    "Share %ldb, Text %ldb, Lib %ldb, Data %ldb\n",
 	    share * page_size, text * page_size, lib * page_size, data * page_size
 	);
-    
-    return;
-    }
-
-/*** Get the current monotonic time in seconds.
- ***
- *** @returns The current monotonic time as a fractional number of seconds.
- ***/
-static double
-get_time(void)
-    {
-    struct timespec ts;
-    
-	clock_gettime(CLOCK_MONOTONIC, &ts);
-    
-    return (double)ts.tv_sec + (double)ts.tv_nsec / 1.0e9f;
-    }
-
-/*** Initialize a timer struct.
- ***
- *** @param timer The timer to initialize.
- *** @returns `timer`, for chaining.
- ***/
-pTimer
-timer_init(pTimer timer)
-    {
-	if (UNLIKELY(timer == NULL)) return NULL;
-	timer->start = NAN;
-	timer->total = 0.0;
-    
-    return timer;
-    }
-
-/*** Allocate and initialize a new timer.
- ***
- *** @returns A newly allocated timer, or NULL if allocation fails.
- ***/
-pTimer
-timer_new(void)
-    {
-    return timer_init(check_ptr(nmMalloc(sizeof(Timer))));
-    }
-
-/*** Start timing.
- ***
- *** @param timer The timer to start.
- *** @returns `timer`, for chaining.
- ***/
-pTimer
-timer_start(pTimer timer)
-    {
-	if (UNLIKELY(timer == NULL)) return NULL;
-	timer->start = get_time();
-    
-    return timer;
-    }
-
-/*** Stop timing and add the elapsed time to the timer total.
- ***
- *** @param timer The timer to stop.
- *** @returns `timer`, for chaining.
- ***/
-pTimer
-timer_stop(pTimer timer)
-    {
-	if (UNLIKELY(timer == NULL)) return NULL;
-	timer->total += get_time() - timer->start;
-    
-    return timer;
-    }
-
-/*** Get the total accumulated time for a timer.
- ***
- *** @param timer The timer to read.
- *** @returns The total accumulated time in seconds, or NAN if `timer`
- *** 	is NULL.
- ***/
-double
-timer_get(pTimer timer)
-    {
-	if (UNLIKELY(timer == NULL)) return NAN;
-    
-    return timer->total;
-    }
-
-/*** Reset a timer to its initial state so that it can be reused to time
- *** something else.
- ***
- *** @param timer The timer to reset.
- *** @returns `timer`, for chaining.
- ***/
-pTimer
-timer_reset(pTimer timer)
-    {
-    return timer_init(timer);
-    }
-
-/*** De-initialize a timer before it is freed.
- ***
- *** @param timer The timer to de-initialize.
- ***/
-void
-timer_de_init(pTimer timer) {}
-
-/*** De-initialize and free a timer allocated by timer_new().
- ***
- *** @param timer The timer to free.
- ***/
-void
-timer_free(pTimer timer)
-    {
-	timer_de_init(timer);
-	nmFree(timer, sizeof(Timer));
     
     return;
     }
